Update 0
I just wrote a whole thing, it took ages - It was great..
Than I updated the Code and I lost it.

It said: Your Great, Watch the moon. and We should release F-16's to Afterburner along the onmicron string. It's basically hair skin and nails that grows out the top of organic life's heads and gets loomed by Aliens. We can take the ships. They seem to be replaceable to the blue Avians 
 me and my dropbears and shoulder mounted fodder cannon can take the Ai/DMT ships. But the string. I wonder if it smells like syinged hair?